+++ base_pack/quac/actions/action_subghz.c
@@ -1,22 +1,17 @@
 // Methods for Sub-GHz transmission
 
 #include <flipper_format/flipper_format_i.h>
-#include <path.h>
-#include <string.h>
 
 #include "helpers/subghz_txrx.h"
 #include <lib/subghz/blocks/custom_btn.h>
-#include <lib/subghz/protocols/raw.h>
 
 #include "action_i.h"
 #include "quac.h"
 
-#define SUBGHZ_DIR_PATH EXT_PATH("subghz/")
-
 typedef struct SubGhzNeedSaveContext {
     App* app;
     SubGhzTxRx* txrx;
-    FuriString* file_path;
+    const FuriString* file_path;
 } SubGhzNeedSaveContext;
 
 void action_subghz_need_save_callback(void* context) {
@@ -46,53 +41,17 @@ void action_subghz_need_save_callback(void* context) {
             break;
         }
     } while(0);
-
-    // Update original .sub file.
-    //In case when rolling code was used in Quac we must update original .sub file with actual rolling code counter
-
-    // Take file name from quac_app path
-    FuriString* quac_filename = furi_string_alloc();
-    furi_string_reset(quac_filename);
-    path_extract_filename(savectx->file_path, quac_filename, false);
-    FURI_LOG_I(TAG, "Extracted quac filename: %s", furi_string_get_cstr(quac_filename));
-
-    //create new char string with full path (dir+filename) to original subghz folder
-    char* full_subghz_file_name =
-        malloc(1 + strlen(SUBGHZ_DIR_PATH) + strlen(furi_string_get_cstr(quac_filename)));
-    strcpy(full_subghz_file_name, SUBGHZ_DIR_PATH);
-    strcat(full_subghz_file_name, furi_string_get_cstr(quac_filename));
-    FURI_LOG_I(TAG, "Full path to safe file: %s", full_subghz_file_name);
-
-    //Save subghz file to original subghz location
-    do {
-        if(!storage_simply_remove(savectx->app->storage, full_subghz_file_name)) {
-            FURI_LOG_E(
-                TAG, "Failed to delete subghz file before re-save in original SUBGHZ location");
-            break;
-        }
-        stream_seek(ff_stream, 0, StreamOffsetFromStart);
-        stream_save_to_file(
-            ff_stream, savectx->app->storage, full_subghz_file_name, FSOM_CREATE_ALWAYS);
-        if(storage_common_stat(savectx->app->storage, full_subghz_file_name, NULL) != FSE_OK) {
-            FURI_LOG_E(
-                TAG, "Error verifying new subghz file after re-save in original SUBGHZ location");
-            break;
-        }
-    } while(0);
-
-    free(full_subghz_file_name);
-    furi_string_free(quac_filename);
 }
 
 static void action_subghz_raw_end_callback(void* context) {
     FURI_LOG_I(TAG, "Stopping TX on RAW");
     furi_assert(context);
-    App* app = context;
+    FuriThread* thread = context;
 
-    app->raw_file_is_tx = false;
+    furi_thread_flags_set(furi_thread_get_id(thread), 0);
 }
 
-void action_subghz_tx(void* context, FuriString* action_path, FuriString* error) {
+void action_subghz_tx(void* context, const FuriString* action_path, FuriString* error) {
     App* app = context;
     const char* file_name = furi_string_get_cstr(action_path);
 
@@ -108,10 +67,9 @@ void action_subghz_tx(void* context, FuriString* action_path, FuriString* error)
 
     FuriString* preset_name = furi_string_alloc();
     FuriString* protocol_name = furi_string_alloc();
+    bool is_raw = false;
 
     subghz_custom_btns_reset();
-
-    app->raw_file_is_tx = false;
 
     FuriString* temp_str;
     temp_str = furi_string_alloc();
@@ -197,6 +155,7 @@ void action_subghz_tx(void* context, FuriString* action_path, FuriString* error)
         if(!strcmp(furi_string_get_cstr(protocol_name), "RAW")) {
             subghz_protocol_raw_gen_fff_data(
                 fff_data, file_name, subghz_txrx_radio_device_get_name(txrx));
+            is_raw = true;
         } else {
             stream_copy_full(
                 flipper_format_get_raw_stream(fff_data_file),
@@ -222,30 +181,18 @@ void action_subghz_tx(void* context, FuriString* action_path, FuriString* error)
         FURI_LOG_E(TAG, "Failed to start TX");
     }
 
-    bool skip_extra_stop = false;
-    FURI_LOG_D(TAG, "Checking if file is RAW...");
-    if(!strcmp(furi_string_get_cstr(protocol_name), "RAW")) {
+    if(is_raw) {
         subghz_txrx_set_raw_file_encoder_worker_callback_end(
-            txrx, action_subghz_raw_end_callback, app);
-        app->raw_file_is_tx = true;
-        skip_extra_stop = true;
-    }
-    do {
-        furi_delay_ms(1);
-    } while(app->raw_file_is_tx);
-
-    if(!app->raw_file_is_tx && !skip_extra_stop) {
-        // TODO: Should this be based on a Setting?
-        furi_delay_ms(1500);
-        subghz_txrx_stop(txrx);
+            txrx, action_subghz_raw_end_callback, furi_thread_get_current());
+        furi_thread_flags_wait(0, FuriFlagWaitAll, FuriWaitForever);
     } else {
-        // TODO: Should this be based on a Setting?
-        furi_delay_ms(50);
-        subghz_txrx_stop(txrx);
+        furi_delay_ms(app->settings.subghz_duration);
     }
-    skip_extra_stop = false;
 
     FURI_LOG_I(TAG, "SUBGHZ: Action complete.");
+
+    // This will call need_save_callback, if necessary
+    subghz_txrx_stop(txrx);
 
     subghz_custom_btns_reset();