+++ base_pack/quac/quac.h
@@ -17,7 +17,7 @@
 #include "item.h"
 
 #define QUAC_NAME    "Quac!"
-#define QUAC_VERSION "v0.8.0"
+#define QUAC_VERSION "v0.9.1"
 #define QUAC_ABOUT                                    \
     "Quick Action remote control\n" QUAC_VERSION "\n" \
     "github.com/rdefeo/quac"
@@ -49,7 +49,6 @@ typedef struct App {
     ItemsView* items_view;
     int depth;
     int selected_item;
-    bool raw_file_is_tx;
 
     FuriString* temp_str; // used for renames/etc
     char temp_cstr[MAX_NAME_LEN]; // used for renames/etc
@@ -59,6 +58,7 @@ typedef struct App {
         QuacAppLayout layout; // Defaults to Portrait
         bool show_icons; // Defaults to True
         bool show_headers; // Defaults to True
+        uint32_t subghz_duration; // Defaults to 1500 ms
         uint32_t rfid_duration; // Defaults to 2500 ms
         uint32_t nfc_duration; // Defaults to 1000 ms
         uint32_t ibutton_duration; // Defaults to 1000 ms