+++ base_pack/quac/scenes/scene_settings.c
@@ -17,7 +17,7 @@
 // dynamically know it's list index for our on_event method. However, we'll need to
 // hardcode the value..
 // TODO: Figure out a better way to do this
-#define SCENE_SETTINGS_ABOUT 8 // 9 items in our Settings list, so last index is 8
+#define SCENE_SETTINGS_ABOUT 9 // 10 items in our Settings list, so last index is 9
 
 static const char* const layout_text[2] = {"Vert", "Horiz"};
 static const uint32_t layout_value[2] = {QUAC_APP_PORTRAIT, QUAC_APP_LANDSCAPE};
@@ -69,6 +69,13 @@ static void scene_settings_show_headers_changed(VariableItem* item) {
     uint8_t index = variable_item_get_current_value_index(item);
     variable_item_set_current_value_text(item, show_offon_text[index]);
     app->settings.show_headers = show_offon_value[index];
+}
+
+static void scene_settings_subghz_duration_changed(VariableItem* item) {
+    App* app = variable_item_get_context(item);
+    uint8_t index = variable_item_get_current_value_index(item);
+    variable_item_set_current_value_text(item, duration_text[index]);
+    app->settings.subghz_duration = duration_value[index];
 }
 
 static void scene_settings_rfid_duration_changed(VariableItem* item) {
@@ -136,6 +143,13 @@ void scene_settings_on_enter(void* context) {
     value_index = value_index_uint32(app->settings.show_headers, show_offon_value, 2);
     variable_item_set_current_value_index(item, value_index);
     variable_item_set_current_value_text(item, show_offon_text[value_index]);
+
+    item = variable_item_list_add(
+        vil, "SubGhz Duration", V_DURATION_COUNT, scene_settings_subghz_duration_changed, app);
+    value_index =
+        value_index_uint32(app->settings.subghz_duration, duration_value, V_DURATION_COUNT);
+    variable_item_set_current_value_index(item, value_index);
+    variable_item_set_current_value_text(item, duration_text[value_index]);
 
     item = variable_item_list_add(
         vil, "RFID Duration", V_DURATION_COUNT, scene_settings_rfid_duration_changed, app);