+++ base_pack/quac/CHANGELOG.md
@@ -1,10 +1,21 @@
-## 0.8.0
+## 0.9.1
+
+- Fixed About screen
+- Updated readme
+
+## 0.9.0
 
 - Complete refactor of SubGhz code
 - Supports dynamic SubGhz signals with rolling codes
 - Auto-detect External SubGhz Antenna and use/prefer if available
 - Removed SugGhz External Antenna setting due to new auto-detect feature
 - Removed SubGhz Repeat setting, no longer needed (was non-standard)
+- Added SubGhz Duration setting for non-RAW signals
+
+## 0.8
+
+- Added Import Link Here, which creates Quac Link files
+- Scroll long action names
 
 ## 0.7.2